move pathname check to header element
